Skill Mode and basic UI
Worked to add a UI and button controls/text fields to the add-on.

Added in a property to hold the skill mode, and changed the rendering setup to accomodate it.

Found a fantastic UI someone had built that incorporates a selector and a window to populate selections. Can be used for all types of things, but planning on using it to grab props and place them into prop lists for each variation.